Avoiding issues with hierarchical namespaces (HNC)
When using hierarchical namespaces, it's possible to configure it
to propagate resources to children namespaces.
It is common to propagate it for Secret, but it can be defined for
any resource. Since trust-manager manage a configmap/secret in every
namespace, the management of the resources can overlap (HNC will prevent
updates on the children namespaces).
It's possible to avoid the resource being propagated by using a custom
label. Let's add it to avoid future conflicts.